Fixed minor security issue. Small refactor.
